requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
#chromadb
langchain_core
langserve[all]
uvicorn
fastapi
streamlit
langchain_groq
# langchain-chroma
langchain_huggingface
sentence_transformers
faiss-cpu
duckdb
pandas
gtts
pdfplumber
